resource-manager: PodStatus-based cgroup discovery.
For Pod discovery (initial sync with the runtime),
fire off an extra PodSandboxStatus request and dig
out the cgroup parent directory from the runtime-
specific info in the reply.

Having to resort to using implementation-specific
data is unfortunate but rather unavoidable at the
moment. In addition to discovering the pod parent
cgroup, we'll soon be using it for recovering the
runtime handler (and the class), which is another
piece of information implementations tend to omit
from the ListPodSandbox response.